pyproject.toml
[tool.poetry]
name = "langgrapg-agent-1"
version = "0.1.0"
description = ""
authors = ["sairakausar <sairakausar3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.62"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
